Add EDIT/UPDATE functionality
Editing was done the harder way. Instead of editing directly on
"Profile page", I redirect to another page for editing. Just wanted to
test and challenge myself.
Also noted when refreshing "editing a post page", firestore data is undefined for a moment(this i need investigation).
Not certain on this, but added an alert box when this happens.
With this, we can catch and redirect unauthorized page access from
direct url

What I learn and applied:
- passing id to route params
- passing props via <Route /> to its component
- react "defaultValue" attr on input fields for setting initial value

Also : fix missed dependancy on previous "useCallback" for delete post